backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
motor==3.6.0
pymongo==4.8.0
redis==5.2.0

# AI & ML
google-generativeai==0.8.3
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.7
langchain-pinecone>=0.2.4
pinecone-client==5.0.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.3.1

# NLP & Text Processing
spacy==3.8.2
textblob==0.18.0
nltk==3.9.1
vaderSentiment==3.3.2

# API & Web
httpx==0.27.2
aiofiles==24.1.0
websockets==13.1
sse-starlette==2.1.3

# Security & Auth
cryptography==44.0.0
argon2-cffi==23.1.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.21.0
sentry-sdk==2.18.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Date & Time
python-dateutil==2.9.0
pytz==2024.2

# Validation & Serialization
marshmallow==3.23.1
email-validator==2.2.0

# Task Queue
celery==5.4.0
kombu==5.4.2

# Rate Limiting
slowapi==0.1.9